requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.0
dnspython>=2.6.1
requests>=2.31.0
geoip2>=5.2.0
maxminddb>=3.0.0
